+++ app/src/main/java/nl/bezorgdirect/mijnbd/api/BodyParameterClasses.kt
@@ -6,7 +6,7 @@ data class UpdateNotificationParams(val accepted: Boolean)
 
 data class UpdateLocationParams(val latitude: Float, val longitude: Float)
 
-data class UpdateStatusParams(val status: Int, val latitude: Float, val longitude: Float)
+data class UpdateStatusParams(val status: Int, val latitude: Double, val longitude: Double)
 
 data class UpdateUserParams (val emailAddress: String?, val phoneNumber: String?, val dateOfBirth: String?, val range: Int?, val vehicle: Int?, val vehicleDisplayName: String?, val fare: Float?, val totalEarnings: Float?)
